backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
anthropic==0.52.0
python-dotenv==1.1.0

# future deps — uncomment when needed
# spacy==3.8.4
# chromadb==1.0.7
# sentence-transformers==4.1.0
# scikit-learn==1.6.1
# faiss-cpu==1.11.0
